# librarySystem_Backend/src/main/java/com/assessment/librarySystem/dto/request/BorrowerRegistrationRequest.java
package com.assessment.librarySystem.dto.request;

import jakarta.validation.constraints.Email;
import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.Size;

/**
 * Request DTO for borrower registration
 */
public record BorrowerRegistrationRequest(
    @NotBlank(message = "Name is required")
    @Size(min = 2, max = 100, message = "Name must be between 2 and 100 characters")
    String name,
    
    @NotBlank(message = "Email is required")
    @Email(message = "Email should be valid")
    String email
) {}